consolidate files and toasts
