+++ context/app/static/js/components/bulkDownload/BulkDownloadButton/index.ts
@@ -1,0 +1,3 @@
+import BulkDownloadButton from './BulkDownloadButton';
+
+export default BulkDownloadButton;
