+++ context/app/static/js/components/bulkDownload/BulkDownloadDialog/index.ts
@@ -1,0 +1,3 @@
+import BulkDownloadDialog from './BulkDownloadDialog';
+
+export default BulkDownloadDialog;
